Remove data files
Data files are now and should continue to be gleaned from pypowervm,
which should be relied upon to keep them current (with respect to the
REST API), clean, and consolidated.

A recent pypowervm change set has added all the data files needed by
this project, so they are removed from here.

The test utilities used to load these files were also refactored (moved
to a different package).  This change set accomodates that refactor.

Change-Id: Ic289c0fda3fe8271465f4c66ab99238d880931b3