Check instance exists on detach_volume
Add a check in the detach_volume() method to see if the instance
is on this host.  In cases where a live migration fails, the
compute manager will attempt to detach each attached volume
from the instance on the destination host, but chances are
the instance is not on the host at all.

Change-Id: I4b3bbd06f653c330f7e52ac7b2190cec2b749e23