Switch from single key to key-per wwpn pairing
In npiv.py, a string mapping of the wwpn pairs are being saved to
system_metadata. In some cases, the string value generated can be
quite large. To prevent this from happening, this change is to move
from a single key to one key per wwpn pairing. We differentiate the
keys by appending a unique number to the end of the key in
_set_fabric_meta. This change requires that _get_fabric_meta also be
modified to accomodate the changes in how the mappings are now stored.

Change-Id: I72e65957edec41405aad30768dab547134b5edce
Closes-Bug: #1492291