Gather physical port on target server
When a live migration occurs, it is important to flip the client WWPNs
as part of the operation.  However, it is also important to update the
system metadata to include the new physical port hosting the WWPNs.

This change set interrogates the physical WWPNs in the
post_live_migration step to determine the correct physical port.  This
is then used to update the system_metadata for the instance.

Change-Id: Icc371d966420cc3dc679db9b70c6a4a4415196b8
Closes-Bug: 1504636