Abort disconnect if invalid udid
The vSCSI physical volume may hit an issue if there is an hdisk without
a UDID.  The connect flow handles this and will not map.  Which then
causes a revert.  However, during the revert the code tries to remove
the hdisk.  This causes a significant issue.

This change set detects this path by adding a flag to check if udid is
None.  If so, it presents a warning message to the operator.  It then
allows the revert flow to continue properly.

Change-Id: If7af35c7e7b68f9a5cf5496fdfded482b45064e0
Closes-Bug: 1505388