Do not map to management partition for NPIV
NPIV based deploys attempted to put the client vFC adapters on to the
fabric by creating the WWPNs ahead of time and mapping them to the
management partition.  This created several issues and led to poor
performance.

There were issues around migration:
- Querying for a host could leave a stale adapter on the mgmt partition

There were issues with speed:
- Deploys could take over 120 seconds
- Migrations had additional delays

There were issues with some standard operations:
- Attaching a volume on a powered off VM required logging in the WWPNs

Subsequent analysis in Cinder showed that only some Cinder drivers had
this requirement.  Many could work without this.  Given the challenges
the code is removing the need to log in the WWPNs on the mgmt partition.
This resolves all of the above issues.

While it does limit some of the Cinder drivers when paired with NPIV,
ongoing work will be done to meet with the Cinder driver owners to see
if the requirements in their drivers can be relaxed.

For those that can not, the vSCSI Cinder Volume attach is also an option
and should provide the use cases that they need.

Change-Id: Id77650545ca4e0def9977f1c2572b9f13f451e0d