Reduce checks for migration wwpn in npiv
The NPIV code would check both the task state and the host attribute to
determine if the WWPN was used for migration.  However, the only check
that needs to be done is the host.  If the instance is not on the new
host, it should use the secondary WWPN.

This change relaxes the code to fit this model.

Change-Id: I314a1a794a47b239b7a882acf327b24f56e1e356
Closes-Bug: 1507018
(cherry picked from commit 23d06863c221d211910edec9d1ee69ac34239fd2)