Initial wwpn only valid if no existing WWPNs
The 'is_initial_wwpn' was simply looking at the fabric state and task
state.  However, if multiple BDMs were specified on a deploy, this would
cause multiple WWPNs to be generated for the same fabric (because each
would be considered 'initial').

This change set expands the logic to also check if any WWPNs have been
set on the fabric yet.  If so, it is not considered an initial WWPN.

Change-Id: I03767f406e004349015483ef8b696f72c57d8ede
Closes-Bug: 1507774