Clear out NPIV metadata on reschedule
When a VM reschedules on to a different host, the NPIV metadata was
sticking around, and this caused an issue with the new host when the VM
landed on it.

This change clears out the system metadata on a full terminate of the
NPIV connection.  A full terminate is when the VM is being destroyed.

This should make the spawn on a new host fully 'fresh'.

Change-Id: I35348bbfffc8cb4693e6d0822e7a0c6ff42de032
Closes-Bug: 1510251