Provide locking around NPIV wwpns method
If the user runs two attach volumes at the same time, this can lead to a
race condition where different WWPNs get generated for the instance.
This change provides a lock around the wwpns call.  This is a global
lock for the compute node.

In previous iterations of this code, it could have had a major
performance impact.  But now that the job is invoked for the WWPNs call,
this method should return very quickly (subsecond).  Therefore the
global lock should not be a concern.

Change-Id: I4100e5c66797eb4f1fad802a88ef90966a05ce13
Closes-Bug: 1514537