Don't remove vOpt mapping on live migrate
Some operating systems are unable to remove a device when it is in use.
However, during a live migrate, we can't migrate the the virtual optical
device.

The current live migrate code was removing the virtual optical media,
but also was removing the mapping.  This change updates it to leave the
mapping (so that operating systems that can't remove the bus are OK),
but to detach the vopt device from it.  We still remove the virtual
optical media.

Change-Id: Id2a3216ec06e95a538fa2f9f25463c490d69f81f
Closes-Bug: 1522222