Handle task state update when expected is None
The image task UpdateTaskState() was unconditionally sending the
expected_state to the update_task_state function.  During a shelve
operation the compute manager uses the expected_state to index
into an array of states.  Passing None causes an exception during
shelving for that reason.  To fix it, we only pass the expected_state
when it's not None.

Change-Id: I00ab265533fe30ba4ffb86759588bba3b139e0a2