Add minor authentication to vnc console
When the VNC console gets created, an 'access path' can be passed to
coordinate the connection.  This helps ensure that the client is
correctly connecting to the appropriate VM.

This change set enables this additional check for the PowerVM driver.
It should not be confused with full authentication.   It simply
validates the UUID.

It is also not encryption.  Other patch sets to the VNC consoles have
been proposed for this.

Change-Id: I3c3bfde9513d9ad96fe5638469a2005ad57e91bb