Implmement NVRAM store and manager
PowerVM has NVRAM that is associated with each VM on the system.
The NVRAM holds information about the VM such as which device
is the boot device. PowerVM produces an event when the NVRAM for
a VM is changed.

The NVRAM should be set on the VM during operations like cold
migration, resize to a different host and evacuation when the
VM is newly created on the destination host from an existing VM
on the source host.

This change set introduces the concept of an NVRAM store and
an NVRAM manager. The compute driver listens for the PowerVM
NVRAM change events and stores the NVRAM in the configured store.

A Swift store for the NVRAM is implemented.

Future change sets will utilize the NVRAM from the store on VM
operations and manage its life cycle. This change set does
include storing the NVRAM during migrate/resize and was included
to facilitate live testing.

Change-Id: I59be9946794f53fdcae5515cc818d4fe967986d0