Conductor changes to register nova-powervm objects
Nova now sends the LiveMigrateData objects exclusively to
pass data between compute managers and conductor during live
migration. Since dicts are no longer supported the conversion
to objects must be done to continue to support live migration.

A previous change set added the object support to the nova-powervm
project.  This change set adds the hack (as the nova core team described
to the PowerVM team on Feb 24th in IRC) that enables operators to use a
custom conductor that supports the PowerVM live migrate object.

Change-Id: I2517d543cc66696a43b2cdb160e2452653f402f1
Closes-Bug: #1551954