Coordinated LU upload 2: marker LUs
Previous change Ie54fd2b205c7ffae90dc960dfe0bfc46f952a8d8 performed
coordinated LU upload to an SSP using an algorithm that assumed it was
possible to rename an existing Logical Unit.  This turns out not to be
the case.

This change set makes a second attempt at coordinated LU upload by
creating a specially-named LU that acts as a marker indicating that the
upload (to the LU with the real, final name) is not yet complete.  The
marker LU is removed when the upload is finished.  The other aspects of
the coordination algorithm are similar:

- Check whether the LU already exists.  (Now this is indicated by the
  presence of the LU with the final name, and NO marker LUs.)  If so,
  return it.
- Check whether an upload is in progress.  (Now this is indicated by the
  presence of any marker LU, whether or not the LU with the real name
  has been created yet.)  If so, spin and wait for it to complete.
- Create a marker LU.
- Pull back the LU list and see if anyone else created a marker and/or
  started the upload while we were creating our marker.
- If someone else started an upload, delete our marker, and spin waiting
  for it to complete.
- If we have multiple markers, alpha-sort them by name.  The first one
  wins.  If it's someone else, delete our marker and spin.
- If there was just one marker (ours), or if ours was first in the
  alpha-sort, do the actual upload.
- Remove the marker LU to signal completion.
- Return the LU we uploaded.

Change-Id: I9c4175532a4f92635c3437ca702aa4d1b4893e50