Use NVRAM data on VM operations
This change set adds support for setting NVRAM data on newly
created VMs during cold migration, evacuation, and unshelving.
The NVRAM is fetched from the NVRAM store and used on the
create of the VM.

Also in this change, the NVRAM data is deleted when the VM is
deleted from the host if it's determined that the data will
not be needed in the future.  Cases where the data should not
be deleted include:

 - Deleting a VM off the source host when it has been cold
   migrated to a destination host.
 - Deleting a VM that has been shelved since the NVRAM data will
   be needed when it's unshelved.
 - Deleting a VM off the source host when it has been evacuated
   to another host.

Change-Id: I11a2bb2d187e2754c2610f3b7ef69f090cf42fb9