Use local partition, not management partition
The PowerVM CI environment is making use of multiple faux 'management
VMs' on a given PowerVM system.  It is doing this to simulate the nested
virtualization that the KVM CI environment uses.  In such cases, the
existing usages of get_mgmt_partition are actually doing the wrong thing
- we instead want to use the partition on which the driver is running.
That turns out to be true regardless - it's just that in all other cases
the local node IS the management partition.

This change set replaces calls to get_mgmt_partition with the new
pypowervm API pypowervm.tasks.partition.get_this_partition.

Also, the get_mgmt_partition method is removed, as it is now available
in pypowervm if it should be needed in the future.

Change-Id: I324ca7a3c3b0bfea426035325296b0a743970fd9