Use any media repository if it exists
A VIOS is only allowed to have one virtual media repository on it.
Attempts to create a second one will fail.

The previous logic in ConfigDrivePowerVM would always try to create a
media repository on the volume group specified via the
vopt_media_volume_group conf option (or rootvg if the option was
omitted).  If a media repository already existed on a different volume
group, this would cause a failure.

With this change set, we look for an existing media repository on any
volume group, on any live VIOS, and use the first one we find.  Only if
we don't find one do we attempt to create it on the configured
vopt_media_volume_group / rootvg.

Change-Id: I33ffa96f95b39aad4de65e3265f120f19021b252
Closes-Bug: 1558253