Cache instance overhead
The inputs to calculate_memory_overhead_on_host are the host UUID, which
is always the same in a NovaLink environment, and a single max_mem value
gleaned from the instance's flavor.  As such, it is not necessary to
invoke the expensive QueryReservedMemoryRequiredForPartition Job every
time we see a given max_mem value.

With this change set, the driver caches the instance overhead value
associated with each max_mem value it sees, making subsequent lookup
against that value effectively instantaneous.

The motivation for this change is scale.  It takes about one second for
QueryReservedMemoryRequiredForPartition to come back.  Since the
estimate_instance_overhead API is invoked via a periodic task, once per
second, serially across all instances, the driver begins to choke once
we reach about 80 instances on the host.

Change-Id: I5650b21b486658921b454a7ac0cd8ab745604b57
Closes-Bug: 1560668