Add live migrate flow framework to vifs
This change set adds the live migrate framework to the VIF drivers.
This allows VIF drivers to take specific actions during a live migration
operation.  These are a no-op for the SEA vif type.  However as other
vif types are added, implementations will be needed.

The obvious need is with the OVS VIF type.  This creates a
'peer-to-peer' connection between the VM and the host.  When a live
migrate occurs, the source trunk adapter is left on the system and will
need to be cleared out.

This changeset is a first step.  Subsequent changes are needed to enable
the OVS flows.

Change-Id: I81940889009b6cb8d7832ea62cb66ffc5adef279