Support roll back of VIF at task level
Previously in the bug, we would only roll back the VIF if something
happened at the OVS level.  But there are many other parts of the
provision pipeline where things can 'go wrong'.

The clean up code was not wrong, it was just a bit narrow.  This change
set integrates those changes up into the TaskFlow pipeline.  That way if
a storage mapping, power on, or some other error occurs in the
provision, the code still calls the unplug.

Change-Id: I09c17b0212b22a5a56b091c9852587d8f91df978
Closes-Bug: 1560466