Added code for deleting orphan trunks
Will search for all trunk adapters without an associated CNA.
It will then delete the ovs port before deleting the trunk adapter.

The trunks on the source host are orphaned after a live migration.
The CNA that was paired with this trunk is now paired with a trunk
on the destination (leaving the source trunk orphaned).
Change-Id: I641543f4010cc6b2bfb013db3a13ec100fc85b34