Allow multiple session connection attempts
We have seen scenarios where users have rebooted their systems and the
ceilometer-powervm process has started before the backing PowerVM REST
server. This generally indicates a packaging issue (where the
ceilometer-powervm package should depend on the others).  However,
there is the ability in the PowerVM Session to make multiple attempts.

This change set takes advantage of the multiple attempts provided by the
pypowervm session.  That means that if the REST server is booting, the
ceilometer-powervm will wait up to 10 minutes for the REST server to
finish starting.  The REST server should start much faster than that,
and this is just an upper limit timeout.

Change-Id: If9569eccbdb9107e5892f03a91a035e1d5943e5e
Closes-Bug: 1570404