Log warning for failed deletion of NVRAM
Currently, if NVRAM deletion fails, an exception is raised and causes
a failure. If we fail to remove the entry from swift, the worst case
scenario is if the data is really in swift, it sticks around and
never gets cleaned out. Instead, if NVRAM deletion fails, we should
log a warning message.

Change-Id: I671f492c2b6640b4a8ce150524d95a09ca5f86b0
Closes-Bug: 1568936