Increase scope of stopable states
Our CI testing has done some hair pin turns on the VMs from create to
destroy.  This uncovered an issue where VMs would be in the 'Shutting
Down' state.  The nova-powervm power off operation would not wait for
the shutting down command to finish.

This change set fixes that issue.

Change-Id: I8a84cf6a9c495ec710aa21ee297b15a08d6f8a21
Closes-Bug: bug/1573042