Support tot_data for metrics
The metrics code was just looking at the deltas between the last thirty
seconds and the new sample.  The cycles it returned to nova were simply
how many cycles were used in that thirty second sample.

Nova needs a running total.  Unfortunately this is a little complicated
in PowerVM as the metrics code keeps the VMs total (no matter which host
it runs on).  That's a weird statement.  To clarify, the base metrics
code will actually move the CPU cycles of a VM onto its new host when it
migrates.  This is different from libvirt which tracks it via the
cycles spent by the process on the host (so post migration, the
'counter' resets to 0).

This means that a VM that spends a million cycles on the source host
will have the cpu cycle counter start at a million when it lands on the
target host.

This change set takes this into account by determining the thirty second
sample delta between the calls.  It then keeps an internal to nova (via
an in memory cache) the total cycles.  This should account for the Live
Migration situation noted above.

Change-Id: Ie03c0dc927bf78f06354cbcbc15565bd3cfd780a
Closes-Bug: 1572820
(cherry picked from commit 52138dfdc404fe96d7d4a159d49e7bad21645671)