Document PowerVM driver functions in support matrix doc
Add a copy of Nova's support matrix document within the docs for
nova-powervm. Within each section a powervm option has been added,
denoting whether or not each feature is currently supported.

This includes the base support matrix files as well as the library
and conf changes required to build it in the nova-powervm docs build.

Change-Id: I46511c61ce35164ec89098c6d876edb26ca8d261