Fix revert migration if flavor deleted
After an instance is created, the flavor could be deleted.  But
operations still need to be able to occur against the VM when that
occurs.  A new test in Tempest showed that we had been using the
incorrect (and potentially deleted) flavor when a resize/migration was
reverting.

This change updates the code to use the flavor that is stored on the
instance itself.

Change-Id: I16254d8eb74a724e3e0a538f47259e9627c4233a
Closes-Bug: 1570748