Add req to pycrypto to fix import errors
We're not sure...

But we think paramiko's recent 2.0.0 release somehow breaks a
requirement chain where *somebody* needs pycrypto and is no longer
getting it.  Adding the requirement for pycrypto (synced with the
corresponding global requirement) resolves the tox errors seen in the
bug report.

And the gate wasn't failing because... the pycrypto req was being picked
up some other way.

Change-Id: I1286b7df7df6608a2fdb929eb55f2d6fbfd031ea
Closes-Bug: 1576822