Allow compute driver to load correctly
In certain environments the load order can be confused between nova and
nova_powervm (now that nova_powervm has a nova/ source path).

This can cause boot issues where 'No module named cmd.compute' can be
thrown because the python code is looking in nova-powervm/nova for that
driver.

This change extends the declare_namespace code to allow for the other
projects to properly load.

This was affecting the PowerVM CI system and not allowing the systems to
properly load.

Change-Id: I4695814ad9ded96a6bb731b2a47f7050fedd8d85