Allow re-mapping of client slots to volumes
When an instance is rebuilt the client slot mappings need to be rebuilt
in the same exact order.  One reason is that the NVRAM will store the
boot list and this boot list depends on the slot mappings.  Also, some
client operating systems also validate the slot configuration against
their database.

This change set handles saving and restoring slot numbers for VSCSI and
NPIV volumes; and for Virtual Ethernet Adapters.

Change-Id: Ib35d5fd4e9a3cad2f1ecfc223d2b8128f97bf86d
Partial-Bug: 1578282