Do not clean NVRAM unless destroy is on same host
When doing a remote rebuild, the source host will likely come back up at
some point and run the cleanup of the evacuated instances.  When this
was occurring, it would clean out the NVRAM and Slot Store information
from Swift (if the operator was using Swift to store such data).

This information should not be wiped out on the source host re-init
because a second remote rebuild operation will need it.

This change updates the logic that identifies when the NVRAM/Slot
information should be deleted.  It adds a check to make sure that the
instance is on either the same host or is set to None.

The existing checks were not enough.  The destroy_disks would get set to
True unless there was shared storage.  Since there may be a rebuild
without shared storage (think boot from Cinder volume), this flag would
get set to true and wipe out the disks.

It also isn't in the KEEP_NVRAM_STATES, because it is not being shelved.
At the same time, extending the KEEP_NVRAM_STATES is a bit of a zero sum
game.  The state is fluid, and there are some states where you may
actually want to destroy the NVRAM, but in other situations in the same
state keep it.

The addition of the host check seems to be most appropriate for this
scenario of multiple remote rebuilds.

Change-Id: I95a824034d657f31285d85ba412a981d6986d14a
Closes-Bug: #1581803