Revert "Removing workaround"
It was identified that there are certain conditions which can cause
issues here and still need this workaround.  It is believed that these
can be replicated with the libvirt driver and bug 1560468 will be
updated to reflect it.

Sridhar Venkat is to own validating that issue with the libvirt driver
and pursuing a global fix.  Until then we will revert this 'removing
workaround' to add it back in.

This reverts commit ea4dcbb3cc163b02e09bc85638fc73f3d8baf12b.

Conflicts:
	nova_powervm/tests/virt/powervm/tasks/test_network.py

Change-Id: I20febda537130da03cac03a75876fda173a99d71