Remove excessive exceptions in NVRAM/SlotStore
The addition of the NVRAM and SlotStore have led to Swift calls to be
made, as that is where the data is stored.  Some of these calls try to
'fetch' the data, and if not there...it is not always a bad thing.
Maybe it is the first deploy?

However, the swiftclient will log errors in this event.  Those errors
start to make 'happy' path code look bad.

This change adds a method called '_exists' to the swift file that is
used in nova_powervm to detect if the NVRAM/SlotStore is actually there.
If it is not there, it will not call the backing download call to Swift.
This prevents a significant amount of logging, but does result in an
extra API call to Swift.

The swiftclient bug #1582347 will address this issue and provide a fix.

Change-Id: Ib6c6f64e845cf94681888a0b5378c95b00b93f25
Closes-Bug: #1581978
(cherry picked from commit 5d801414378e2120b633ae3ee5bb5115aba40517)