Check instance.host on NPIV disconnect
During source host evacuation cleanup, _destroy_evacuated_instances,
the manager invokes driver destroys for these instances. The instances
never have their task state changed so the NPIV driver's _disconnect
never removes the mappings.

Add a check to see if the instance.host is the same as this host
or if it's None for a spawn. If it's not then we can cleanup the
mappings.

Closes-Bug: #1582850
Change-Id: I0464cef0b73879f98559e96fdc9a0863f4fcbcb9