Add LUA to remote restart/rebuild VSCSI mappings
When rebuilding an LPAR and preserving slot assignments for I/O devices,
it is also important to preserve the logical unit address (LUA) of each
storage device associated with a VSCSI mapping.  Client operating
systems rely on these LUAs for such things as proper boot device
ordering and device attachment validation.

This change set pulls the LUA along with the slot number from the
rebuild map and passes it along to the creation of the new VSCSI mapping
on the RR target side.  (On the first build, the behavior is similar to
other lookups: None will be returned for the LUA informing the server to
determine the LUA for the device.)

Change-Id: I5b4346f5effc3fd1d6eb4226cf1068c9e535ddb7
Partial-Bug: 1578282