vif.unplug is invoked with a missing argument during rollback
When a vif create failure occurs, transaction is rolled back with
revert method in PlugVifs in tasks/network.py. This method invokes
vif.unplug with one less argument - slot_mgr.
The same call in execute method in UnplugVifs is called with all
required parameters.
This fix adds missing parameter to the call and related unit tests.

Change-Id: I93b6204ab8426d390832520e1d5af6a798648ba4
Closes-Bug: 1583211