Scrub before RebuildSlotMap
The spawn flow adds scrubbing tasks for the VM being created, to be run
after VIF creation, but before any storage mappings occur.

On a rebuild, VIF plugging needs to generate (and validate) the slot
mappings for the I/O devices being attached to the rebuilding VM. The
process of creating the slot mappings includes ensuring all the
necessary disks are present and available on the target VIOSes. Since
this is happening before the scrubbing, it can fail when stale mappings
still exist for the VM being created.

This change set adds a comprehensive scrub in
NovaSlotManager.init_recreate_map, before the RebuildSlotMap is
initialized, which includes scrubbing all mappings from stale LPARs
*and* the LPAR we're rebuilding.

Change-Id: I87bb279141a2e867f5349ce7e699596420b1c0dc
Closes-Bug: 1583825