Reuse existing connection
The swift operations were rebuilding the swift client with every
operation.  This is very expensive, calls over to keystone often and is
generally very slow.

This change supports building a swift service that is used across many
operations.  This should speed up operations quite a bit, especially
with a reboot of the system.

Change-Id: Iacbc87f26cf2e568bc662a96b6d9b459b814c98f
Partial-Bug: 1584946