Handle stale system metadata (physical WWPNs)
In certain scenarios, the system metadata for the NPIV fabric
configurations may no longer be valid.  This can occur when one adds a
volume, disconnects the volume, migrates the VM, and then adds a volume
on the new host.

When this happens, the following occurs.
 - First add: generates NPIV port map (physical WWPN to virtual WWPNs)
 - Remove of volume: does nothing.  Port maps are cleaned out only on
   delete of VM.
 - Migrate
 - Second add: The old maps point to the old physical WWPNs.

One solution would be to clear out the system metadata on the detach.
But since one NPIV mapping exists for all the volumes, one would need to
track if this is the 'last volume to be removed'.  That is very
challenging.

This change set instead detects on the add volume if the metadata is
incorrect.  It will then rebuild the metadata accordingly and update the
physical mappings. The rebuild needs to reset the fabric state and
handle the stale mapping (no backing_port) in pypowervm. So a secondary
change is needed there too.

Change-Id: I853e4a7f5dcb5e09c550a2c4756794a45c1bf8dd
Partial-Bug: #1585624