Raise an exception if no active VIOSes in build_tx_feed_task
get_active_vioses returns an empty list if there aren't any
active VIOSes. If an empty list is used as the feed to a FeedTask
then we get a generic error explaining that the FeedTask can't have
an empty feed.

This change attempts to make the error more clear in the event
there aren't any active VIOSes when build_tx_feed_task is called
by raising an exception explaining that there aren't any active
VIOSes.

Change-Id: I3d67c653131952bfc13bbbda2e67486c0291c4cc
Closes-Bug: #1581652