Update the PHYP vSwitch for Neutron Agents
There is a PHYP virtual switch that is used when the PowerVM Nova driver
is paired with the Neutron Open vSwitch or Linux Bridge agent.
Initially this was named 'OpenStackOVS', but this change set renames it
to the 'NovaLinkVEABridge' to bring it more in line with its general
purpose nature.

Change-Id: Ib4521f2286ba7963c3a40303f6013c4b5a6522f7