Some VM deploys fail from expired keystone token
Deploy of a virtual machine occasionally fails.  The failure occurs
about every 6 hours, as the issue is tied to Keystone's token
renewal.  The NVRAM upload operation using Swift client maybe using
an expired keystone token which is causing the operation to fail with
a ClientException error. Enabling the regeneration of the token
without the exception by retrying the upload operation after a
ClientException error is caught.

Change-Id: I2c89d4d6c85370e3daaef11d8be1a24b12ed2ca4
Closes-Bug: #1594833