Save valid udid on pre_live_migration of vscsi volumes using multi VIOS
The pre_live_migration_on_destination operation for vscsi volumes saved
the udid to be used later during cleanup. In cases where the target
contains multiple VIOS and if one of them do not contain valid ports
then the mig_data will set with udid as None.

Adding code to set the udid only if it is not None.

Change-Id: I3412663a6dab828bbfa2f7fa1c938aa956b7b315
Closes-Bug: 1598393