Remove delete vopt if nothing to delete
This change set optimizes the delete of the virtual optical a bit.
If the user has choosen not to use config drive (which is what uses
vopt) the delete of the media would still call an update to the media
repo.  This wasn't that expensive but didn't need to be called in this
scenario.

Change-Id: Ic9e7ba2aad7887ffaeec727e799996c91d593c5a