Provide logs for each task start
This is meant to help with debugging CI runs.  Indicates quickly how
long each task took to run.

This change set also removes some now redundant messages in the tasks.
Lastly, it also updates some of the task names to be more clear as
they're now represented in the logs directly.

Change-Id: I6ca776f1651d4c9542da0561e4152c40f1e4553a