Plug updated VIFs
A plug of a VIF does two operations.  It creates the client side
adapter.  But it also does a 'treat' of the VIF so that the neutron
agent is aware of it.

It is expected that when the nova compute boots up that all of the VIFs
get plugged.  Specifically so that they can be retreated.

As a performance optimization, the PowerVM driver did not run the
plug_vifs if we detected that the client network adapter was already
there.  However, this led to an issue.  The VIF wasn't being
re-'treated' on the nova compute boot.

This wasn't an issue, unless you had the I/O running through the VM that
the nova-compute was running on (like OVS or Linux Bridge).  If that
VM were to be rebooted, then the elements on the trunk adapter (like the
OVS UUID) are lost.  They do not persist via reboot.

This is the OpenStack design and specifically why OpenStack calls
plug_vifs on each compute node reboot.  This change brings the
nova-powervm driver in line with that behavior.

The biggest change is that the OVS will now set the UUID on the trunk
port and therefore the OVS neutron agent will finish its initialization.

Change-Id: I1fcd708a39f87282a77a8c8db75dcbbff83e4d4e
Closes-Bug: #1609497